trellis: separate independent unit test, introduce sane test numbering
